+++ src/MXNet.jl
@@ -34,6 +34,20 @@ export NDArray,
        context,
        empty
 
+# autograd.jl
+export attach_grad!,
+       backward!,
+       getgrad,
+       is_recording,
+       is_training,
+       mark_variables,
+       pause,
+       predict_mode,
+       record,
+       symbol,
+       train_mode,
+       @custom
+
 # executor.jl
 export Executor,
        bind,